# src/main/java/ru/extoozy/iotapplication/api/dto/LoginResponseDto.java
package ru.extoozy.iotapplication.api.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class LoginResponseDto {

    private String access;
    private String refresh;
}